Move SubtitleExtractor to text package
PiperOrigin-RevId: 399400909
