+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/text/SubtitleExtractor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.exoplayer2.extractor.subtitle;
+package com.google.android.exoplayer2.extractor.text;
 
 import static com.google.android.exoplayer2.util.Assertions.checkState;
 import static com.google.android.exoplayer2.util.Assertions.checkStateNotNull;
