+++ library/extractor/src/test/java/com/google/android/exoplayer2/extractor/text/SubtitleExtractorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.exoplayer2.extractor.subtitle;
+package com.google.android.exoplayer2.extractor.text;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
